+++ CHANGELOG.md
@@ -7,8 +7,17 @@ Versioned according to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 Added
+
+  * find_files supports `ID`
+
 Changed
+
+  * Adding a file with an existing `ID` will raise an exception, #110
+
 Fixed
+
+  * `GROUPID` was not passed when adding files
+  
 Removed
 
 ## [0.2.5] - 2018-06-18